Safari-611.1.10/LayoutTests/http/wpt/credential-management/credentialscontainer-preventSilentAccess-basics.https.html

<!DOCTYPE html>
<title>Credential Management API: preventSilentAccess() basics.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(function(t) {
    return promise_rejects_dom(t, "NotSupportedError",
            navigator.credentials.preventSilentAccess());
}, "navigator.credentials.preventSilentAccess().");
</script>
